+++ src/easy_language_learning_tool/__init__.py
@@ -1,3 +1,3 @@
 """Easy Language Learning Tool."""
 
-__version__ = "1.4.1"
+__version__ = "1.5.0"